# questao_4/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API de Academia de Ginástica",
    "description": "Essa é a documentação da API de Academia de Ginástica desenvolvida para o teste prático da Niky",
    "version": "1.0.0",
    "contact": {
      "email": "henrikoalberton@gmail.com"
    }
  },
  "paths": {
...
    "/matriculas": {
      "post": {
        "tags": ["Matrículas"],
        "summary": "Cria uma Matrícula",
        "description": "Cria uma nova Matrícula",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Matricula"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Matrícula criada com sucesso"
          }
        }
      },
      "get": {
        "tags": ["Matrículas"],
        "summary": "Lista as Matrículas",
        "description": "Lista todas as Matrículas",
        "responses": {
          "200": {
            "description": "Retorna todas as Matrículas"
          }
        }
      }
    },
    "/matriculas/{id}": {
      "get": {
        "tags": ["Matrículas"],
        "summary": "Lista uma Matrícula",
        "description": "Lista os detalhes de uma Matrícula",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da Matrícula",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna os detalhes de uma Matrícula"
          }
        }
      },
      "put": {
        "tags": ["Matrículas"],
        "summary": "Atualiza uma Matrícula",
        "description": "Atualiza os dados de uma Matrícula",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da Matrícula",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Matricula"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matrícula atualizada com sucesso"
          }
        }
      },
      "delete": {
        "tags": ["Matrículas"],
        "summary": "Deleta uma Matrícula",
        "description": "Deleta uma Matrícula",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da Matrícula",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Matrícula deletada com sucesso"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Matricula": {
      "type": "object",
      "properties": {
        "aluno": { "type": "number", "example": 1, "required": true },
        "turma": { "type": "number", "example": 1, "required": true }
      }
    },
...
  }
}